Refactor user ratings and post into api module file
